requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.9.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==3.0.1
attrdict==2.0.1
attrs==21.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bitarray==3.4.2
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==4.2.2
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.4
ckzg==2.1.1
click==8.2.1
cloudflare==2.8.15
colorama==0.4.6
Cython==3.1.2
cytoolz==1.0.1
dateparser==1.1.0
docutils==0.15.2
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fake-useragent==0.1.11
Flask==3.1.1
Flask-APScheduler==1.12.2
Flask-Mail==0.9.1
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==2.8
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.9.4
jsonlines==2.0.0
jsonschema==3.2.0
Mako==1.3.10
MarkupSafe==3.0.2
multiaddr==0.0.9
multidict==6.4.4
netaddr==0.8.0
noraise==0.0.16
orjson==3.10.18
parsimonious==0.10.0
pdfid==1.1.3
pika==1.2.0
propcache==0.3.2
protobuf==3.17.3
psycopg2-binary==2.9.10
pycparser==2.19
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==1.7.1
pyrsistent==0.20.0
python-binance==1.0.29
python-bitcoinrpc==1.0
python-dateutil==2.8.0
python-dotenv==1.1.0
python-editor==1.0.4
python-telegram-bot==13.8.1
pytz==2019.1
pytz-deprecation-shim==0.1.0.post0
pyunormalize==16.0.0
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.26.0
rfc3986==1.5.0
rlp==4.1.0
s3transfer==0.13.0
setuptools==80.9.0
simple-multiprocessing==0.0.15
simplejson==3.16.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.2.1
SQLAlchemy==1.3.24
stopit==1.1.2
toolz==1.0.0
tornado==6.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2021.5
tzlocal==4.1
ujson==5.10.0
urllib3==1.26.20
varint==1.0.2
web3==7.12.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
